Fix string test bug with temporary reference
In C++, string literals are const char*, not std::string's. Since adding
support for raw array serialization, a more suitable overload was found
for the Save function, meaning "hello" gets serialized as an array of
characters, omitting the length. This was causing this test to fail, but
is correct behavior. The test has now been updated to use a std::string
temporary, instead of char*.